implemented options for UseKissLogMiddleware extension #11
