+++ src/KissLog.AspNetCore/IOptionsBuilder.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace KissLog.AspNetCore
+{
+    public interface IOptionsBuilder
+    {
+        List<ILogListener> Listeners { get; }
+        Options Options { get; }
+    }
+}
